Create rename_columns
In many situations, changing column names may be a bad choice. However, when your goal is simply to analyze data or generate visualizations, using a DataFrame with columns names containing whitespace, accents or symbols can be quite annoying.

It’s easier to select columns with standardized names, not to mention when someone left that unwanted whitespace at the end of an Excel column name.