Create generate_fake_date.py
Whether for testing or learning purposes, we often need to create fake variables. For example, you might need fake data to write unit tests or to play with a new library.

Of course, we can always import or create these variables from scratch, but there is a much easier and more versatile solution!