Create dataframe_eda.py
There are some basic analysis that needs to be done on any imported/generated DataFrame. How about creating a function to perform these repetitive steps and just importing it *(as shown in the second trick)*

I’ve already developed this function for my personal use, and I thought it would be cool to make it available here for anyone who wants to use, modify, or have as a basis to create their own!